Dot and PNG
added dot and png for actual and pointer
