+++ prog1/LocationManagementAlgorithms.py
@@ -9,327 +9,15 @@ __date__ = '$Date: 2016/02/06 12:00:00 $'
 __copyright__ = 'Copyright (c) 2016 James Soehlke and David N. Taylor'
 __license__ = 'Python'
 
-POINTER = "pointer"
-VALUE = "value"
-FORWARDING_P = "forwarding_p"
-FORWARDING_V = "forwarding_v"
-FORWARDING_P_FORWARD = "forwarding_p_forward"
-FORWARDING_P_REVERSE = "forwarding_p_reverse"
-MOBILE_STATION = "mobile station"
-LOCATION = "location"
-REPLICATION = "replication"
-REPLICATION_P = "replication_p"
-REPLICATION_V = "replication_v"
-
-algorithm_list = [POINTER, VALUE, FORWARDING_P, FORWARDING_V, REPLICATION_P, REPLICATION_V]
-location_list = [FORWARDING_P_FORWARD, FORWARDING_P_REVERSE, REPLICATION]
-location_list[len(location_list):] = algorithm_list
-
-
-class Tree(object):
-    def __init__(self, algorithm):
-        self.__root_node = None
-        self.__leaf_nodes = []
-        self.__algorithm = algorithm
-        if self.__algorithm is not None:
-            self.__algorithm.set_tree(self)
-        self.__node_update_list = []
-        self.__node_search_list = []
-        return
-
-    def add_node(self, node):
-        if node.is_root():
-            self.__root_node = node
-        elif node.is_leaf():
-            self.__leaf_nodes.append(node)
-        return
-
-    def get_root_node(self):
-        return self.__root_node
-
-    def get_leaf_nodes(self):
-        return self.__leaf_nodes
-
-    def get_node(self, name):
-        if self.__root_node is not None:
-            return self.__root_node.get_child_with_name(name)
-
-        return None
-
-    @staticmethod
-    def get_lca(node1, node2):
-        n1 = node1
-        n2 = node2
-
-        if (n1 is None) or (n2 is None):
-            return None
-
-        while n1 is not n2:
-            while not n2.is_root():
-                n2 = n2.get_parent_node()
-                if n1 is n2:
-                    return n2
-
-            if n1.is_root():
-                break
-            else:
-                n1 = n1.get_parent_node()
-                n2 = node2
-
-        if n1 is n2:
-            return n2
-        else:
-            return None
-
-    def put_ms_into_node_name(self, ms, name):
-        node = self.get_node(name)
-        if node is None:
-            return False
-
-        if (self.__algorithm.get_type() == VALUE) or (self.__algorithm.get_type() == FORWARDING_V) or (self.__algorithm.get_type() == REPLICATION_V):
-            node.add_ms_location(ms, node, self.__algorithm.get_type())
-        else:
-            node.add_ms_location(ms, node, self.__algorithm.get_type())
-        ms.set_node(node, self.__algorithm.get_type())
-
-        #print node.get_name()
-
-        while not node.is_root():
-            node_parent = node.get_parent_node()
-            if (self.__algorithm.get_type() == VALUE) or (self.__algorithm.get_type() == FORWARDING_V) or (self.__algorithm.get_type() == REPLICATION_V):
-                node_parent.add_ms_location(ms, ms.get_node(self.__algorithm.get_type()), self.__algorithm.get_type())
-            else:
-                node_parent.add_ms_location(ms, node, self.__algorithm.get_type())
-            node = node_parent
-            #print node.get_name()
-
-        return True
-
-    def get_update_count(self):
-        return self.__algorithm.get_update_count()
-
-    def get_search_count(self):
-        return self.__algorithm.get_search_count()
-
-    def find_node_and_query_ms_location_from_node(self, ms, name):
-        del self.__node_search_list[:]
-        return self.__algorithm.find_node_and_query_ms_location_from_node(ms, name)
-
-    def query_ms_location_from_node(self, ms, node):
-        del self.__node_search_list[:]
-        return self.__algorithm.query_ms_location_from_node(self, ms, node)
-
-    def find_node_and_move_ms_location_from_node(self, ms, name):
-        del self.__node_update_list[:]
-        return self.__algorithm.find_node_and_move_ms_location_from_node(ms, name)
-
-    def move_ms_to_node(self, ms, node):
-        del self.__node_update_list[:]
-        return self.__algorithm.move_ms_to_node(ms, node)
-
-    def get_algorithm(self):
-        return self.__algorithm
-
-    def add_node_search(self, node):
-        if self.__node_search_list.count(node) == 0:
-            self.__node_search_list.append(node)
-
-    def add_node_update(self, node):
-        if not self.has_node_been_updated(node):
-            self.__node_update_list.append(node)
-
-    def has_node_been_updated(self, node):
-        return self.__node_update_list.count(node) > 0
-
-    def draw_tree(self):
-        return
-
-
-class Node(object):
-    def __init__(self, name):
-        self.__name = name
-        self.__parent_node = None
-        self.__children_node = []
-        self.__ms_list = {}
-        self.__lcmr = 0.0
-        return
-
-    def get_name(self):
-        return self.__name
-
-    def get_children_nodes(self):
-        return self.__children_node
-
-    def get_child_with_name(self, name):
-        if self.get_name() == name:
-            return self
-
-        for i in range(len(self.__children_node)):
-            n = self.__children_node[i].get_child_with_name(name)
-            if n is not None:
-                return n
-
-        return None
-
-    def add_child_node(self, node):
-        if self.__children_node.count(node) == 0:
-            self.__children_node.append(node)
-        return
-
-    def add_parent_node(self, parent):
-        self.__parent_node = parent
-        return
-
-    def get_parent_node(self):
-        return self.__parent_node
-
-    def find_ms(self, ms, loc=None):
-        it = iter(self.__ms_list)
-        for i in it:
-            if ms is self.__ms_list[i][MOBILE_STATION]:
-                if loc is not None:
-                    if self.__ms_list[i][LOCATION][loc] is not None:
-                        return i
-                else:
-                    return i
-
-        return None
-
-    def get_new_ms_list_index(self):
-        i = 0
-        it = iter(self.__ms_list)
-        for j in it:
-            if i <= j:
-                i = j + 1
-        return i
-
-    def add_ms(self, ms):
-        if self.find_ms(ms) is None:
-            l = self.get_new_ms_list_index()
-            self.__ms_list[l] = {}
-            self.__ms_list[l][MOBILE_STATION] = ms
-            self.__ms_list[l][LOCATION] = {}
-            it = iter(location_list)
-            for loc in it:
-                self.__ms_list[l][LOCATION][loc] = None
-            #print self.__ms_list.__len__()
-            #print self.__ms_list
-        return
-
-    def contains_ms(self, ms):
-        return self.find_ms(ms) is None
-
-    def is_root(self):
-        return self.__parent_node is None
-
-    def is_leaf(self):
-        return len(self.__children_node) == 0
-
-    def add_ms_location(self, ms, loc, loc_type):
-        if self.find_ms(ms) is None:
-            self.add_ms(ms)
-            i = self.find_ms(ms)
-            self.__ms_list[i][LOCATION][loc_type] = loc
-            #print self.__ms_list
-            return True
-        else:
-            return self.set_ms_location(ms, loc, loc_type)
-
-    def set_ms_location(self, ms, loc, loc_type):
-        i = self.find_ms(ms)
-        if i is not None:
-            self.__ms_list[i][LOCATION][loc_type] = loc
-            #print self.__ms_list
-            return True
-
-        return False
-
-    def get_ms_location(self, ms, loc_type):
-        i = self.find_ms(ms)
-        if i is not None:
-            return self.__ms_list[i][LOCATION][loc_type]
-
-        return None
-
-    def delete_ms_location(self, ms, loc_type):
-        i = self.find_ms(ms)
-        if i is not None:
-            self.__ms_list[i][LOCATION][loc_type] = None
-            empty = True
-            it = iter(location_list)
-            for loc in it:
-                if self.__ms_list[i][LOCATION][loc] is not None:
-                    empty = False
-            if empty:
-                del self.__ms_list[i]
-                #print self.__ms_list
-                return True
-            else:
-                return False
-        return True
-
-    def delete_all_ms_locations(self, ms):
-        i = self.find_ms(ms)
-        if i is not None:
-            del self.__ms_list[i]
-            #print self.__ms_list
-        return
-
-    def set_lcmr(self, lcmr):
-        self.__lcmr = lcmr
-        return
-
-    def get_lcmr(self):
-        return self.__lcmr
-
-    def calculate_lcmr(self):
-        if self.is_leaf():
-            return self.__lcmr
-
-        self.__lcmr = 0
-
-        it = iter(self.__children_node)
-        for i in it:
-            self.__lcmr += i.calculate_lcmr()
-
-        return self.__lcmr
-
-
-
-class MS(object):
-    def __init__(self, name, max_forwards=5):
-        self.__name = name
-        self.__node_list = {}
-        self.__max_forwards = max_forwards
-        it = iter(algorithm_list)
-        for node_type in it:
-            self.__node_list[node_type] = None
-        return
-
-    def set_ms_name(self, name):
-        self.__name = name
-        return
-
-    def get_ms_name(self):
-        return self.__name
-
-    def set_node(self, node, loc_type):
-        self.__node_list[loc_type] = node
-        return
-
-    def get_node(self, loc_type):
-        return self.__node_list[loc_type]
-
-    def get_max_forwards(self):
-        return self.__max_forwards
-
+from LocationManagementMisc import *
 
 class Algorithm(object):
-    def __init__(self, algorithm_type):
+    def __init__(self, algorithm_type, name=""):
         self.__search_count = 0
         self.__update_count = 0
         self.__tree = None
         self.__type = algorithm_type
+        self.__name = name
         return
 
     def set_search_count(self, value):
@@ -364,6 +52,9 @@ class Algorithm(object):
     def get_type(self):
         return self.__type
 
+    def get_name(self):
+        return self.__name
+
     def find_node_and_query_ms_location_from_node(self, ms, name):
         return
 
@@ -376,10 +67,13 @@ class Algorithm(object):
     def move_ms_to_node(self, ms, node):
         return None
 
+    def print_dot_search(self, f, search_list):
+        return
+
 
 class ValueAlgorithm(Algorithm):
-    def __init__(self, loc_type=VALUE):
-        Algorithm.__init__(self, loc_type)
+    def __init__(self, loc_type=VALUE, name="Actual_Value_Algorithm"):
+        Algorithm.__init__(self, loc_type, name)
         return
 
     def find_node_and_query_ms_location_from_node(self, ms, name):
@@ -440,26 +134,37 @@ class ValueAlgorithm(Algorithm):
 
         while lca is not ms_node:
             ms_node.delete_ms_location(ms, self.get_type())
+            self.get_tree().add_node_update(ms_node)
             ms_node = ms_node.get_parent_node()
             self.increment_update_count()
-            self.get_tree().add_node_update(ms_node)
 
         ms_node = node
 
         while ms_node is not None:
             ms_node_parent = ms_node.get_parent_node()
             ms_node.add_ms_location(ms, node, self.get_type())
+            self.get_tree().add_node_update(ms_node)
             ms_node = ms_node.get_parent_node()
             self.increment_update_count()
-            self.get_tree().add_node_update(ms_node)
 
         ms.set_node(node, self.get_type())
         return
 
+    def print_dot_search(self, f, search_list):
+        if len(search_list) > 1:
+            for i in range(len(search_list)):
+                j = i + 1
+                if j >= len(search_list):
+                    return
+                text = str(search_list[i].get_name()) + " -- " + str(search_list[j].get_name())
+                text = text + "[style=bold,color=red,label=\"S" + str(i) + "\"];"
+                f.write(text)
+                f.write("\n")
+
 
 class PointerAlgorithm(Algorithm):
-    def __init__(self, loc_type=POINTER):
-        Algorithm.__init__(self, loc_type)
+    def __init__(self, loc_type=POINTER, name="Pointer_Algorithm"):
+        Algorithm.__init__(self, loc_type, name)
         return
 
     def find_node_and_query_ms_location_from_node(self, ms, name):
@@ -507,8 +212,8 @@ class PointerAlgorithm(Algorithm):
                 return node
 
             while node is not node.get_ms_location(ms, self.get_type()):
-                self.get_tree().add_node_search(node)
                 node = node.get_ms_location(ms, self.get_type())
+                self.get_tree().add_node_search(node)
                 self.increment_search_count()
 
             return node
@@ -532,9 +237,9 @@ class PointerAlgorithm(Algorithm):
 
         while lca is not ms_node:
             ms_node.delete_ms_location(ms, self.get_type())
+            self.get_tree().add_node_update(ms_node)
             ms_node = ms_node.get_parent_node()
             self.increment_update_count()
-            self.get_tree().add_node_update(ms_node)
 
         ms_node = node
 
@@ -549,20 +254,27 @@ class PointerAlgorithm(Algorithm):
         self.increment_update_count()
         self.get_tree().add_node_update(lca)
 
-        # if lca is not ms_node:
-        #     ms_node.add_ms_location(ms, ms_node, self.get_type())
-        #     self.increment_update_count()
-
         node.add_ms_location(ms, node, self.get_type())
         self.increment_update_count()
-        self.get_tree().add_node_update(ms_node)
+        self.get_tree().add_node_update(node)
         ms.set_node(node, self.get_type())
         return
+
+    def print_dot_search(self, f, search_list):
+        if len(search_list) > 1:
+            for i in range(len(search_list)):
+                j = i + 1
+                if j >= len(search_list):
+                    return
+                text = str(search_list[i].get_name()) + " -- " + str(search_list[j].get_name())
+                text = text + "[style=bold,color=red,label=\"S" + str(i) + "\"];"
+                f.write(text)
+                f.write("\n")
 
 
 class ForwardingPointerPAlgorithm(PointerAlgorithm):
     def __init__(self, max_forwards=2):
-        PointerAlgorithm.__init__(self, FORWARDING_P)
+        PointerAlgorithm.__init__(self, FORWARDING_P, "Forwarding_Algorithm_For_Pointer")
         self.__forwarding_type = FORWARDING_P_FORWARD
         self.__reverse_type = FORWARDING_P_REVERSE
         self.__max_forwards = max_forwards
@@ -787,9 +499,12 @@ class ForwardingPointerPAlgorithm(PointerAlgorithm):
 
         return node
 
+    def print_dot_search(self, f, search_list):
+        return
+
 class ForwardingPointerVAlgorithm(ValueAlgorithm):
     def __init__(self, max_forwards=2):
-        ValueAlgorithm.__init__(self, FORWARDING_V)
+        ValueAlgorithm.__init__(self, FORWARDING_V, "Forwarding_Algorithm_For_Actual_Value")
         self.__forwarding_type = FORWARDING_P_FORWARD
         self.__reverse_type = FORWARDING_P_REVERSE
         self.__max_forwards = max_forwards
@@ -983,7 +698,7 @@ class ReplicationValues(object):
 
 class ReplicationPointerAlgorithm(PointerAlgorithm, ReplicationValues):
     def __init__(self):
-        PointerAlgorithm.__init__(self, REPLICATION_P)
+        PointerAlgorithm.__init__(self, REPLICATION_P, "Replication_Algorithm_For_Pointer")
         ReplicationValues.__init__(self)
         return
     
@@ -1074,10 +789,13 @@ class ReplicationPointerAlgorithm(PointerAlgorithm, ReplicationValues):
                 node = node.get_parent_node()
         print "Rcount is %d" % rcount
 
+    def print_dot_search(self, f, search_list):
+        return
+
 
 class ReplicationValueAlgorithm(ValueAlgorithm, ReplicationValues):
     def __init__(self):
-        ValueAlgorithm.__init__(self, REPLICATION_V)
+        ValueAlgorithm.__init__(self, REPLICATION_V, "Forwarding_Algorithm_For_Actual_Value")
         ReplicationValues.__init__(self)
         return
     
@@ -1152,190 +870,5 @@ class ReplicationValueAlgorithm(ValueAlgorithm, ReplicationValues):
                 node = node.get_parent_node()
         print "Rcount is %d" % rcount
 
-
-def fill_tree(tree):
-    node0 = Node(0)
-    node1 = Node(1)
-    node2 = Node(2)
-    node3 = Node(3)
-    node4 = Node(4)
-    node5 = Node(5)
-    node6 = Node(6)
-    node7 = Node(7)
-    node8 = Node(8)
-    node9 = Node(9)
-    node10 = Node(10)
-    node11= Node(11)
-    node12 = Node(12)
-    node13 = Node(13)
-    node14 = Node(14)
-    node15 = Node(15)
-    node16 = Node(16)
-    node17 = Node(17)
-    node18 = Node(18)
-
-    node7.set_lcmr(.2)
-    node8.set_lcmr(.2)
-    node9.set_lcmr(.2)
-    node10.set_lcmr(.2)
-    node11.set_lcmr(.2)
-    node12.set_lcmr(.2)
-    node13.set_lcmr(.2)
-    node14.set_lcmr(.2)
-    node15.set_lcmr(.2)
-    node16.set_lcmr(.2)
-    node17.set_lcmr(.2)
-    node18.set_lcmr(.2)
-
-    node0.add_child_node(node1)
-    node0.add_child_node(node2)
-    node1.add_parent_node(node0)
-    node2.add_parent_node(node0)
-
-    node1.add_child_node(node3)
-    node1.add_child_node(node4)
-    node3.add_parent_node(node1)
-    node4.add_parent_node(node1)
-
-    node2.add_child_node(node5)
-    node2.add_child_node(node6)
-    node5.add_parent_node(node2)
-    node6.add_parent_node(node2)
-
-    node3.add_child_node(node7)
-    node3.add_child_node(node8)
-    node3.add_child_node(node9)
-    node7.add_parent_node(node3)
-    node8.add_parent_node(node3)
-    node9.add_parent_node(node3)
-
-    node4.add_child_node(node10)
-    node4.add_child_node(node11)
-    node4.add_child_node(node12)
-    node10.add_parent_node(node4)
-    node11.add_parent_node(node4)
-    node12.add_parent_node(node4)
-
-    node5.add_child_node(node13)
-    node5.add_child_node(node14)
-    node5.add_child_node(node15)
-    node13.add_parent_node(node5)
-    node14.add_parent_node(node5)
-    node15.add_parent_node(node5)
-
-    node6.add_child_node(node16)
-    node6.add_child_node(node17)
-    node6.add_child_node(node18)
-    node16.add_parent_node(node6)
-    node17.add_parent_node(node6)
-    node18.add_parent_node(node6)
-
-    tree.add_node(node0)
-    for i in range(18):
-        tree.add_node(tree.get_node(i))
-
-    return
-
-
-if __name__ == "__main__":
-    tpa = Tree(PointerAlgorithm())
-    tfppa = Tree(ForwardingPointerPAlgorithm())
-    tva = Tree(ValueAlgorithm())
-    tfpva = Tree(ForwardingPointerVAlgorithm())
-    trva = Tree(ReplicationValueAlgorithm())
-    trpa = Tree(ReplicationPointerAlgorithm())
-    fill_tree(tpa)
-    fill_tree(tva)
-    fill_tree(tfppa)
-    fill_tree(tfpva)
-    fill_tree(trva)
-    fill_tree(trpa)
-
-    ms1 = MS(1)
-
-    tpa.put_ms_into_node_name(ms1, 8)
-    tva.put_ms_into_node_name(ms1, 8)
-    tfppa.put_ms_into_node_name(ms1, 8)
-    tfpva.put_ms_into_node_name(ms1, 8)
-    trpa.put_ms_into_node_name(ms1, 8)
-    trva.put_ms_into_node_name(ms1, 8)
-
-    ms2 = MS(2)
-
-    tpa.put_ms_into_node_name(ms2, 17)
-    tva.put_ms_into_node_name(ms2, 17)
-    tfppa.put_ms_into_node_name(ms2, 17)
-    tfpva.put_ms_into_node_name(ms2, 17)
-    trpa.put_ms_into_node_name(ms2, 17)
-    trva.put_ms_into_node_name(ms2, 17)
-
-    #tpa.find_node_and_move_ms_location_from_node(ms1, 13)
-    #print "PA Update Count  = %d" % tpa.get_update_count()
-    tpa.find_node_and_move_ms_location_from_node(ms1, 13)
-    p = tpa.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(POINTER).get_name())
-    #tva.find_node_and_move_ms_location_from_node(ms1, 13)
-    #print "VA Update Count  = %d" % tva.get_update_count()
-    tva.find_node_and_move_ms_location_from_node(ms1, 13)
-    v = tva.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(VALUE).get_name())
-    #tfppa.find_node_and_move_ms_location_from_node(ms1, 13)
-    #print "FPP Update Count = %d" % tfppa.get_update_count()
-    tfppa.find_node_and_move_ms_location_from_node(ms1, 11)
-    fpp = tfppa.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(FORWARDING_P).get_name())
-    print "FPP Search Count = %d" % tfppa.get_search_count()
-    print "FPP Update Count = %d" % tfppa.get_update_count()
-    tfppa.find_node_and_move_ms_location_from_node(ms1, 13)
-    fpp = tfppa.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(FORWARDING_P).get_name())
-    print "FPP Search Count = %d" % tfppa.get_search_count()
-    print "FPP Update Count = %d" % tfppa.get_update_count()
-    tfppa.find_node_and_move_ms_location_from_node(ms1, 15)
-    fpp = tfppa.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(FORWARDING_P).get_name())
-    print "FPP Search Count = %d" % tfppa.get_search_count()
-    print "FPP Update Count = %d" % tfppa.get_update_count()
-    tfppa.find_node_and_move_ms_location_from_node(ms1, 13)
-    fpp = tfppa.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(FORWARDING_P).get_name())
-    #tfpva.find_node_and_move_ms_location_from_node(ms1, 13)
-    #print "FPV Update Count = %d" % tfpva.get_update_count()
-    #tfpva.find_node_and_move_ms_location_from_node(ms1, 11)
-    #fpv = tfpva.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(FORWARDING_V).get_name())
-    #print "FPV Search Count = %d" % tfpva.get_search_count()
-    #print "FPV Update Count = %d" % tfpva.get_update_count()
-    #tfpva.find_node_and_move_ms_location_from_node(ms1, 13)
-    #fpv = tfpva.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(FORWARDING_V).get_name())
-    #print "FPV Search Count = %d" % tfpva.get_search_count()
-    #print "FPV Update Count = %d" % tfpva.get_update_count()
-    #tfpva.find_node_and_move_ms_location_from_node(ms1, 12)
-    #fpv = tfpva.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(FORWARDING_V).get_name())
-    #print "FPV Search Count = %d" % tfpva.get_search_count()
-    #print "FPV Update Count = %d" % tfpva.get_update_count()
-    tfpva.find_node_and_move_ms_location_from_node(ms1, 13)
-    fpv = tfpva.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(FORWARDING_V).get_name())
-    trpa.find_node_and_move_ms_location_from_node(ms1, 13)
-    p = trpa.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(POINTER).get_name())
-    trva.find_node_and_move_ms_location_from_node(ms1, 13)
-    p = trva.find_node_and_query_ms_location_from_node(ms1, ms2.get_node(POINTER).get_name())
-    print
-    print
-    print "PA Search Count  = %d" % tpa.get_search_count()
-    print "PA Update Count  = %d" % tpa.get_update_count()
-    print "VA Search Count  = %d" % tva.get_search_count()
-    print "VA Update Count  = %d" % tva.get_update_count()
-    print "FPP Search Count = %d" % tfppa.get_search_count()
-    print "FPP Update Count = %d" % tfppa.get_update_count()
-    print "FPV Search Count = %d" % tfpva.get_search_count()
-    print "FPV Update Count = %d" % tfpva.get_update_count()
-    print "RPA Search Count  = %d" % trpa.get_search_count()
-    print "RPA Update Count  = %d" % trpa.get_update_count()
-    print "RVA Search Count  = %d" % trva.get_search_count()
-    print "RVA Update Count  = %d" % trva.get_update_count()
-    print
-    print
-
-    if (p.get_name() == v.get_name()) and (p.get_name() == fpp.get_name()) and (p.get_name() == fpv.get_name()):
-        print "All search results are the same."
-    else:
-        print "All search results are NOT the same!!"
-        print "p    = %d" % p.get_name()
-        print "v    = %d" % v.get_name()
-        print "fpp  = %d" % fpp.get_name()
-        print "fpv  = %d" % fpv.get_name()
-
+    def print_dot_search(self, f, search_list):
+        return
