+++ prog1/LocationManagementApp.py
@@ -43,6 +43,7 @@ class LocationManagementLabels:
 
     def __initialize(self):
         self.__addHeaderLabel()
+        self.__addLabel("Algorithm Option")
         self.__addLabel("Leaf Node 7 CMR")
         self.__addLabel("Leaf Node 8 CMR")
         self.__addLabel("Leaf Node 9 CMR")
@@ -75,6 +76,13 @@ class LocationManagemenEntry:
         self.__row += 1
         return self.__row
 
+    def __addDropDownEntry(self, varName, setFocus=False):
+        self.__inputs[varName] = Tkinter.StringVar()
+        self.__inputs[varName].set("Pointer")
+        self.__entries[varName] = Tkinter.OptionMenu(self.__parent,self.__inputs[varName], 'Pointer','Actual','FPPointer', 'FPActual', 'RPointer', 'RActual')
+        self.__entries[varName].grid(column = self.__col, row = self.__nextRow(), sticky='EW')
+        #self.__entries[varName].pack()
+
     def __addDoubleEntry(self, varName, initValue=0, setFocus=False):
         self.__inputs[varName] = Tkinter.StringVar()
         self.__inputs[varName].set(format(initValue,".2f"))
@@ -96,6 +104,7 @@ class LocationManagemenEntry:
     def __initialize(self):
         self.__header = Tkinter.Label(self.__parent, text = "Values", width = 20)
         self.__header.grid(column = self.__col, row = self.__nextRow(), sticky = "W")
+        self.__addDropDownEntry("algo_option", self.__setFocus)
         self.__addDoubleEntry("node7cmr", .3, self.__setFocus)
         self.__addDoubleEntry("node8cmr", .3, self.__setFocus)
         self.__addDoubleEntry("node9cmr", .3, self.__setFocus)
@@ -110,7 +119,7 @@ class LocationManagemenEntry:
         self.__addDoubleEntry("node18cmr", .3, self.__setFocus)
         self.__addDoubleEntry("minS", .6, self.__setFocus)
         self.__addDoubleEntry("maxS", 1.2, self.__setFocus)
-        self.__addIntEntry("ms1movetonode", 1)
+        self.__addIntEntry("ms1movetonode", 7)
         self.__addIntEntry("ms2movetonode", 18)
 
         self.__calcButton = Tkinter.Button(self.__parent,text="Move MS1", command=self.__OnMS1MoveClick)
@@ -136,7 +145,7 @@ class LocationManagemenEntry:
 
     def __OnMS1MoveClick(self):
         print "__OnMS1MoveClick"
-        self.__parent.move_ms1(int(self.__inputs["ms1movetonode"].get()))
+        self.__parent.move_ms1(int(self.__inputs["ms1movetonode"].get()), self.__inputs)
 
     def __OnMS2MoveClickEvt(self, evt):
         print "__OnMS2MoveClickEvt"
@@ -144,7 +153,7 @@ class LocationManagemenEntry:
 
     def __OnMS2MoveClick(self):
         print "__OnMS2MoveClick"
-        self.__parent.move_ms2(int(self.__inputs["ms2movetonode"].get()))
+        self.__parent.move_ms2(int(self.__inputs["ms2movetonode"].get()), self.__inputs)
 
     def __OnMS2CallMS1MoveClickEvt(self, evt):
         print "__OnMS2CallMS1MoveClickEvt"
@@ -232,16 +241,88 @@ class LocationManagementApp(Tkinter.Tk):
         self.__calcs.append(LocationManagemenEntry(self, self.__calcNames[0][1][0].get(), i+2, -1, i==0))
         self.update()
 
-    def move_ms1(self, name):
-        print "move ms1 to %d" % name
+    def update_trees(self, inputs):
+        self.__trpa.get_node(7).set_lcmr(float(inputs["node7cmr"].get()))
+        self.__trpa.get_node(8).set_lcmr(float(inputs["node8cmr"].get()))
+        self.__trpa.get_node(9).set_lcmr(float(inputs["node9cmr"].get()))
+        self.__trpa.get_node(10).set_lcmr(float(inputs["node10cmr"].get()))
+        self.__trpa.get_node(11).set_lcmr(float(inputs["node11cmr"].get()))
+        self.__trpa.get_node(12).set_lcmr(float(inputs["node12cmr"].get()))
+        self.__trpa.get_node(13).set_lcmr(float(inputs["node13cmr"].get()))
+        self.__trpa.get_node(14).set_lcmr(float(inputs["node14cmr"].get()))
+        self.__trpa.get_node(15).set_lcmr(float(inputs["node15cmr"].get()))
+        self.__trpa.get_node(16).set_lcmr(float(inputs["node16cmr"].get()))
+        self.__trpa.get_node(17).set_lcmr(float(inputs["node17cmr"].get()))
+        self.__trpa.get_node(18).set_lcmr(float(inputs["node18cmr"].get()))
+        self.__trpa.get_algorithm().set_mins(float(inputs["minS"].get()))
+        self.__trpa.get_algorithm().set_mins(float(inputs["maxS"].get()))
+        self.__trva.get_node(7).set_lcmr(float(inputs["node7cmr"].get()))
+        self.__trva.get_node(8).set_lcmr(float(inputs["node8cmr"].get()))
+        self.__trva.get_node(9).set_lcmr(float(inputs["node9cmr"].get()))
+        self.__trva.get_node(10).set_lcmr(float(inputs["node10cmr"].get()))
+        self.__trva.get_node(11).set_lcmr(float(inputs["node11cmr"].get()))
+        self.__trva.get_node(12).set_lcmr(float(inputs["node12cmr"].get()))
+        self.__trva.get_node(13).set_lcmr(float(inputs["node13cmr"].get()))
+        self.__trva.get_node(14).set_lcmr(float(inputs["node14cmr"].get()))
+        self.__trva.get_node(15).set_lcmr(float(inputs["node15cmr"].get()))
+        self.__trva.get_node(16).set_lcmr(float(inputs["node16cmr"].get()))
+        self.__trva.get_node(17).set_lcmr(float(inputs["node17cmr"].get()))
+        self.__trva.get_node(18).set_lcmr(float(inputs["node18cmr"].get()))
+        self.__trva.get_algorithm().set_mins(float(inputs["minS"].get()))
+        self.__trva.get_algorithm().set_mins(float(inputs["maxS"].get()))
 
-    def move_ms2(self, name):
+    def move_ms1(self, name, inputs):
+        print "move ms1 to %d" % name
+        self.update_trees(inputs)
+        if inputs["algo_option"].get() == "Pointer":
+            self.__tpa.find_node_and_move_ms_location_from_node(self.__ms1, name)
+        elif inputs["algo_option"].get() == "Actual":
+            self.__tva.find_node_and_move_ms_location_from_node(self.__ms1, name)
+        elif inputs["algo_option"].get() == "FPPointer":
+            self.__tfppa.find_node_and_move_ms_location_from_node(self.__ms1, name)
+        elif inputs["algo_option"].get() == "FPActual":
+            self.__tfpva.find_node_and_move_ms_location_from_node(self.__ms1, name)
+        elif inputs["algo_option"].get() == "RPointer":
+            self.__trpa.find_node_and_move_ms_location_from_node(self.__ms1, name)
+        elif inputs["algo_option"].get() == "RActual":
+            self.__trva.find_node_and_move_ms_location_from_node(self.__ms1, name)
+
+    def move_ms2(self, name, inputs):
         print "move ms2 to %d" % name
+        self.update_trees(inputs)
+        if inputs["algo_option"].get() == "Pointer":
+            self.__tpa.find_node_and_move_ms_location_from_node(self.__ms2, name)
+        elif inputs["algo_option"].get() == "Actual":
+            self.__tva.find_node_and_move_ms_location_from_node(self.__ms2, name)
+        elif inputs["algo_option"].get() == "FPPointer":
+            self.__tfppa.find_node_and_move_ms_location_from_node(self.__ms2, name)
+        elif inputs["algo_option"].get() == "FPActual":
+            self.__tfpva.find_node_and_move_ms_location_from_node(self.__ms2, name)
+        elif inputs["algo_option"].get() == "RPointer":
+            self.__trpa.find_node_and_move_ms_location_from_node(self.__ms2, name)
+        elif inputs["algo_option"].get() == "RActual":
+            self.__trva.find_node_and_move_ms_location_from_node(self.__ms2, name)
 
     def ms2_call_ms1(self, inputs):
         print "ms2 call ms1"
-        for k in inputs:
-            print inputs[k].get()
+        if inputs["algo_option"].get() == "Pointer":
+            self.__tpa.query_ms_location_from_node(self.__ms1, self.__ms2.get_node(self.__tpa.get_algorithm().get_type()))
+            self.__tpa.draw_tree(self.__ms1, self.__ms2)
+        elif inputs["algo_option"].get() == "Actual":
+            self.__tva.query_ms_location_from_node(self.__ms1, self.__ms2.get_node(self.__tva.get_algorithm().get_type()))
+            self.__tva.draw_tree(self.__ms1, self.__ms2)
+        elif inputs["algo_option"].get() == "FPPointer":
+            self.__tfppa.query_ms_location_from_node(self.__ms1, self.__ms2.get_node(self.__tfppa.get_algorithm().get_type()))
+            self.__tfppa.draw_tree(self.__ms1, self.__ms2)
+        elif inputs["algo_option"].get() == "FPActual":
+            self.__tfpva.query_ms_location_from_node(self.__ms1, self.__ms2.get_node(self.__tfpva.get_algorithm().get_type()))
+            self.__tfpva.draw_tree(self.__ms1, self.__ms2)
+        elif inputs["algo_option"].get() == "RPointer":
+            self.__trpa.query_ms_location_from_node(self.__ms1, self.__ms2.get_node(self.__trpa.get_algorithm().get_type()))
+            self.__trpa.draw_tree(self.__ms1, self.__ms2)
+        elif inputs["algo_option"].get() == "RActual":
+            self.__trva.query_ms_location_from_node(self.__ms1, self.__ms2.get_node(self.__trva.get_algorithm().get_type()))
+            self.__trva.draw_tree(self.__ms1, self.__ms2)
 
 if __name__ == "__main__":
     app = LocationManagementApp(None)
